Fixes for Debian4 x32 nested VM

+++ scripts/nested-vm.sh
@@ -67,14 +67,23 @@ fi
 # Install KVM and other tools.
 sudo apt -qy update
 sudo apt -qy install rsync
-# Since Debian 9, single libvirt-bin was split into two
-# Note that we don't run this script on other platforms but Debian 8 and 9,
-# so we don't need to care about, say, Ubuntu VERSION=18.04
-if grep -q VERSION.*8 /etc/os-release
-then
-    sudo apt -qy install libvirt-bin
-else
-    sudo apt -qy install libvirt-daemon-system libvirt-clients
+# Since Debian 9 and Ubuntu 18, single libvirt-bin was split into two.
+# This script supports running on Debian 8 and 9, and Ubuntu 16 and 18
+if grep -q 'ID=ubuntu' /etc/os-release
+then # Ubuntu
+    if grep -q VERSION.*16 /etc/os-release
+    then
+        sudo apt -qy install libvirt-bin qemu-utils qemu qemu-kvm
+    else
+        sudo apt -qy install libvirt-daemon-system libvirt-clients
+    fi
+else # Debian
+    if grep -q VERSION.*8 /etc/os-release
+    then
+        sudo apt -qy install libvirt-bin
+    else
+        sudo apt -qy install libvirt-daemon-system libvirt-clients
+    fi
 fi
 
 # Enable nbd devices to have partitions.
@@ -92,6 +101,7 @@ fi
 # Mount the image and add some keys.
 sudo qemu-nbd -d /dev/nbd0 || true
 sudo qemu-nbd -c /dev/nbd0 $DISK
+sudo partprobe -s /dev/nbd0
 sudo mkdir -p /mnt/vm
 success=0
 for i in 1 2 3 4 5
@@ -155,23 +165,30 @@ fi
 sed -i -e "s,[^']*/$BASEDISK,$DISK," $XML
 
 chmod go+rx $HOME
-sudo chown libvirt-qemu:libvirt-qemu $DISK $XML
+libvirt_group="$(grep libvirt /etc/group | cut -d: -f1 | head -n1)"
+sudo chown libvirt-qemu:$libvirt_group $DISK $XML
 
 # Start the VM
 sudo virsh net-start default || true
-# Follow the Google guide:
-# https://cloud.google.com/compute/docs/instances/enable-nested-virtualization-vm-instances
-sudo apt-get -y install uml-utilities qemu-kvm bridge-utils tmux
-sudo modprobe dummy
-sudo brctl delif virbr0 dummy0 || true
-sudo brctl addif virbr0 dummy0
-sudo tunctl -t tap0 || true
-sudo ifconfig tap0 up
-sudo brctl delif virbr0 tap0 || true
-sudo brctl addif virbr0 tap0
-MAC=`sed -e '/mac address/!d' -e "s/.*'\(.*\)'.*/\1/" $XML`
-sudo pkill qemu-system-x86_64 || true
-tmux new-session -d "sudo qemu-system-x86_64 -enable-kvm -hda $DISK -m 789 -curses -netdev tap,ifname=tap0,script=no,id=hostnet0 -device rtl8139,netdev=hostnet0,id=net0,mac=$MAC,bus=pci.0,addr=0x3"
+if sudo dmesg | grep -q "BIOS Google"
+then
+    # We're in Google Cloud, so follow the Google guide:
+    # https://cloud.google.com/compute/docs/instances/enable-nested-virtualization-vm-instances
+    sudo apt-get -y install uml-utilities qemu-kvm bridge-utils tmux
+    sudo modprobe dummy
+    sudo brctl delif virbr0 dummy0 || true
+    sudo brctl addif virbr0 dummy0
+    sudo tunctl -t tap0 || true
+    sudo ifconfig tap0 up
+    sudo brctl delif virbr0 tap0 || true
+    sudo brctl addif virbr0 tap0
+    MAC=`sed -e '/mac address/!d' -e "s/.*'\(.*\)'.*/\1/" $XML`
+    sudo pkill qemu-system-x86_64 || true
+    tmux new-session -d "sudo qemu-system-x86_64 -enable-kvm -hda $DISK -m 789 -curses -netdev tap,ifname=tap0,script=no,id=hostnet0 -device rtl8139,netdev=hostnet0,id=net0,mac=$MAC,bus=pci.0,addr=0x3"
+else
+    # do it like we did before
+    sudo virsh create $XML
+fi
 
 # WAIT for host and find its IP
 IP=
